Migrate Homepage (#71)
Migrated the homepage over to docs-next.

As the default VuePress theme is not flexible enough to support the current homepage's layout (most notably the jumbotron/hero area), I've created a (minimal) custom theme instead. This theme also supports HTML in the tagline, feature blocks, and footer, allowing us to have e.g. newlines and links in these elements. I've also gone ahead and moved some custom components like SocialIcon down (up?) to the theme level.

The result will not be a 1:1 copycat of current vuejs.org UI, with the most significant difference being the fonts. We'll need to come back and revise this once a new theme for Vue 3 docs is available. Also, a mocked Patreon sponsor list is being used after a discussion with Evan. Other functionalities should be in place.