Clarify IME caveat tip (#804)
Previously, the tip on using `input` to listen to IME-based composition input was unclear. In the context of the preceding statement about `textarea` and `input type=text` elements using `input`, it seemed that composition with v-model should work with those inputs. This change makes it clear that `v-model` should be replaced with a custom `@input/:value` pair.